refactoring types to enable DRY business logic
