+++ src/Components/Types/SidebarPropType.tsx
@@ -1,0 +1,6 @@
+type SidebarPropType = {
+  drawerOpen: boolean;
+  toggleFilter: Function;
+};
+
+export default SidebarPropType;
